% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_price_table.R
\name{fetch_price_table}
\alias{fetch_price_table}
\title{Download and extract EC2 pricing table from web page}
\usage{
fetch_price_table()
}
\description{
Download and extract EC2 pricing table from web page
}
